Remove create webhook page and cleanups
