--- src/webhooks/components/WebhookCreatePage/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./WebhookCreatePage";
-export * from "./WebhookCreatePage";
