+++ src/webhooks/components/WebhookDetailsPage/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./WebhookDetailsPage";
+export * from "./WebhookDetailsPage";
