--- src/webhooks/components/WebhooksDetailsPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./WebhooksDetailsPage";
-export * from "./WebhooksDetailsPage";
